requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
msgpack==1.1.0
PyJWT==2.9.0
redis==5.2.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
